+++ src/Recruit/Transport/Controller/Api/V1/Application/ApplicationStatusUpdateController.php
@@ -4,11 +4,9 @@ declare(strict_types=1);
 
 namespace App\Recruit\Transport\Controller\Api\V1\Application;
 
-use App\Recruit\Application\Service\ApplicationDiscussionBootstrapService;
-use App\Recruit\Domain\Enum\ApplicationStatus;
+use App\Recruit\Application\Service\ApplicationStatusTransitionService;
 use App\Recruit\Infrastructure\Repository\ApplicationRepository;
 use App\User\Domain\Entity\User;
-use DomainException;
 use OpenApi\Attributes as OA;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -19,11 +17,6 @@ use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Security\Core\Authorization\Voter\AuthenticatedVoter;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 
-use function array_key_exists;
-use function in_array;
-use function is_string;
-use function strtoupper;
-
 #[AsController]
 #[OA\Tag(name: 'Recruit Application')]
 #[IsGranted(AuthenticatedVoter::IS_AUTHENTICATED_FULLY)]
@@ -31,7 +24,7 @@ class ApplicationStatusUpdateController
 {
     public function __construct(
         private readonly ApplicationRepository $applicationRepository,
-        private readonly ApplicationDiscussionBootstrapService $applicationDiscussionBootstrapService,
+        private readonly ApplicationStatusTransitionService $applicationStatusTransitionService,
     ) {
     }
 
@@ -67,55 +60,13 @@ class ApplicationStatusUpdateController
 
         /** @var array<string, mixed> $payload */
         $payload = $request->toArray();
-        $status = $payload['status'] ?? null;
+        $this->applicationStatusTransitionService->applyStatusTransition($application, $payload['status'] ?? null);
 
-        if (!is_string($status)) {
-            throw new HttpException(JsonResponse::HTTP_BAD_REQUEST, 'Field "status" must be provided as a string.');
-        }
-
-        $newStatus = ApplicationStatus::tryFrom(strtoupper($status));
-        if ($newStatus === null) {
-            throw new HttpException(JsonResponse::HTTP_BAD_REQUEST, 'Field "status" must be one of: WAITING, IN_PROGRESS, DISCUSSION, INVITE_TO_INTERVIEW, INTERVIEW, ACCEPTED, REJECTED.');
-        }
-
-        $currentStatus = $application->getStatus();
-        if ($newStatus !== $currentStatus && !$this->isAllowedTransition($currentStatus, $newStatus)) {
-            throw new HttpException(JsonResponse::HTTP_BAD_REQUEST, 'Status transition is not allowed for this application.');
-        }
-
-        if ($newStatus === ApplicationStatus::DISCUSSION && $currentStatus !== ApplicationStatus::DISCUSSION) {
-            try {
-                $this->applicationDiscussionBootstrapService->bootstrap($application);
-            } catch (DomainException $exception) {
-                throw new HttpException(JsonResponse::HTTP_UNPROCESSABLE_ENTITY, $exception->getMessage(), $exception);
-            }
-        }
-
-        $application->setStatus($newStatus);
         $this->applicationRepository->save($application);
 
         return new JsonResponse([
             'id' => $application->getId(),
             'status' => $application->getStatusValue(),
         ]);
-    }
-
-    private function isAllowedTransition(ApplicationStatus $from, ApplicationStatus $to): bool
-    {
-        $allowedTransitions = [
-            ApplicationStatus::WAITING->value => [ApplicationStatus::IN_PROGRESS->value, ApplicationStatus::REJECTED->value],
-            ApplicationStatus::IN_PROGRESS->value => [ApplicationStatus::DISCUSSION->value, ApplicationStatus::INVITE_TO_INTERVIEW->value, ApplicationStatus::REJECTED->value],
-            ApplicationStatus::DISCUSSION->value => [ApplicationStatus::INVITE_TO_INTERVIEW->value, ApplicationStatus::REJECTED->value],
-            ApplicationStatus::INVITE_TO_INTERVIEW->value => [ApplicationStatus::INTERVIEW->value, ApplicationStatus::REJECTED->value],
-            ApplicationStatus::INTERVIEW->value => [ApplicationStatus::ACCEPTED->value, ApplicationStatus::REJECTED->value],
-            ApplicationStatus::ACCEPTED->value => [],
-            ApplicationStatus::REJECTED->value => [],
-        ];
-
-        if (!array_key_exists($from->value, $allowedTransitions)) {
-            return false;
-        }
-
-        return in_array($to->value, $allowedTransitions[$from->value], true);
     }
 }
