+++ src/Recruit/Transport/Controller/Api/V1/Job/JobCreateFromApplicationController.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace App\Recruit\Transport\Controller\Api\V1\Job;
 
 use App\General\Application\Message\EntityCreated;
+use App\Recruit\Application\Service\JobPayloadHydratorService;
+use App\Recruit\Application\Service\RecruitResolverService;
 use App\Recruit\Domain\Entity\Job;
-use App\Recruit\Domain\Entity\Recruit;
-use App\Recruit\Domain\Repository\Interfaces\RecruitRepositoryInterface;
 use App\Recruit\Infrastructure\Repository\JobRepository;
 use App\User\Domain\Entity\User;
 use OpenApi\Attributes as OA;
@@ -28,12 +28,11 @@ use function trim;
 #[IsGranted(AuthenticatedVoter::IS_AUTHENTICATED_FULLY)]
 class JobCreateFromApplicationController
 {
-    use JobPayloadHydratorTrait;
-
     public function __construct(
-        private readonly RecruitRepositoryInterface $recruitRepository,
+        private readonly RecruitResolverService $recruitResolverService,
         private readonly JobRepository $jobRepository,
         private readonly MessageBusInterface $messageBus,
+        private readonly JobPayloadHydratorService $jobPayloadHydratorService,
     ) {
     }
 
@@ -48,7 +47,7 @@ class JobCreateFromApplicationController
             throw new HttpException(JsonResponse::HTTP_BAD_REQUEST, 'Field "title" is required and must be a non-empty string.');
         }
 
-        $recruit = $this->resolveRecruitByApplicationSlug($applicationSlug);
+        $recruit = $this->recruitResolverService->resolveByApplicationSlug($applicationSlug);
         $application = $recruit->getApplication();
 
         if ($application?->getUser()?->getId() !== $loggedInUser->getId()) {
@@ -59,7 +58,7 @@ class JobCreateFromApplicationController
             ->setRecruit($recruit)
             ->setTitle(trim($title));
 
-        $this->applyJobFields($job, $payload);
+        $this->jobPayloadHydratorService->applyJobFields($job, $payload);
 
         $this->jobRepository->save($job);
         $this->messageBus->dispatch(new EntityCreated('recruit_job', $job->getId(), context: [
@@ -73,16 +72,5 @@ class JobCreateFromApplicationController
             'slug' => $job->getSlug(),
             'title' => $job->getTitle(),
         ], JsonResponse::HTTP_CREATED);
-    }
-
-    private function resolveRecruitByApplicationSlug(string $applicationSlug): Recruit
-    {
-        $recruit = $this->recruitRepository->findOneByApplicationSlug($applicationSlug);
-
-        if (!$recruit instanceof Recruit) {
-            throw new HttpException(JsonResponse::HTTP_BAD_REQUEST, 'Unknown "applicationSlug".');
-        }
-
-        return $recruit;
     }
 }
