+++ src/Recruit/Transport/Controller/Api/V1/Job/PrivateJobStatsController.php
@@ -5,8 +5,7 @@ declare(strict_types=1);
 namespace App\Recruit\Transport\Controller\Api\V1\Job;
 
 use App\Recruit\Application\Service\JobStatsService;
-use App\Recruit\Domain\Entity\Recruit;
-use App\Recruit\Domain\Repository\Interfaces\RecruitRepositoryInterface;
+use App\Recruit\Application\Service\RecruitResolverService;
 use App\User\Domain\Entity\User;
 use OpenApi\Attributes as OA;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -23,7 +22,7 @@ use Symfony\Component\Security\Http\Attribute\IsGranted;
 class PrivateJobStatsController
 {
     public function __construct(
-        private readonly RecruitRepositoryInterface $recruitRepository,
+        private readonly RecruitResolverService $recruitResolverService,
         private readonly JobStatsService $jobStatsService,
     ) {
     }
@@ -31,23 +30,12 @@ class PrivateJobStatsController
     #[Route(path: '/v1/recruit/private/{applicationSlug}/jobs/stats', methods: [Request::METHOD_GET])]
     public function __invoke(string $applicationSlug, User $loggedInUser): JsonResponse
     {
-        $recruit = $this->resolveRecruitByApplicationSlug($applicationSlug);
+        $recruit = $this->recruitResolverService->resolveByApplicationSlug($applicationSlug);
 
         if ($recruit->getApplication()?->getUser()?->getId() !== $loggedInUser->getId()) {
             throw new HttpException(JsonResponse::HTTP_FORBIDDEN, 'You cannot access stats for this application.');
         }
 
         return new JsonResponse($this->jobStatsService->getStats($recruit));
-    }
-
-    private function resolveRecruitByApplicationSlug(string $applicationSlug): Recruit
-    {
-        $recruit = $this->recruitRepository->findOneByApplicationSlug($applicationSlug);
-
-        if (!$recruit instanceof Recruit) {
-            throw new HttpException(JsonResponse::HTTP_BAD_REQUEST, 'Unknown "applicationSlug".');
-        }
-
-        return $recruit;
     }
 }
